+++ NEWS.md
@@ -1,6 +1,7 @@
 # nrvtools (development version)
 
 - make explicit the dependency on R >= 4.1 due to use of native pipe (`|>`);
+- remove dependency package `raster` (#2);
 - move `qs` package to Suggests;
 
 # nrvtools 0.0.21